+++ static_src/components/messageField.jsx
@@ -5,7 +5,7 @@ import SendIcon from 'material-ui/svg-icons/content/send';
 import Message from './message';
 import { bindActionCreators } from "redux";
 import connect from "react-redux/es/connect/connect";
-
+import { sendMessage } from "../actions/messageActions";
 
 class MessageField extends React.Component {
     static propTypes = {
@@ -24,7 +24,11 @@ class MessageField extends React.Component {
 
     sendMessage = (message, sender) => {
         if (this.state.input.length > 0 || sender === 'bot') {
-            this.props.sendMessage(message, sender);
+            const { chatId, messages } = this.props;
+            const messageId = Object.keys(messages).length + 1;
+
+            // Вызываем Action-метод Redux
+            this.props.sendMessage(messageId, message, sender, chatId);
         }
 
         if (sender === 'me') {
@@ -45,6 +49,17 @@ class MessageField extends React.Component {
     // Ставим фокус на <input> при монтировании компонента
     componentDidMount() {
         this.textInput.current.focus();
+    };
+
+    componentDidUpdate(prevProps, prevState) {
+        const { messages } = this.props;
+
+        if (Object.values(messages)[Object.values(messages).length - 1].sender === 'me') {
+            setTimeout( 
+                () => this.sendMessage('Не приставай ко мне, я робот!', 'bot'),
+                1000 
+            );
+        }
     };
 
     render() {
@@ -83,11 +98,12 @@ class MessageField extends React.Component {
     }
 }
 
-const mapStateToProps = ({ chatReducer }) => ({
+const mapStateToProps = ({ chatReducer, messageReducer }) => ({
     chats: chatReducer.chats,
+    messages: messageReducer.messages,
  });
  
- const mapDispatchToProps  = dispatch => bindActionCreators({}, dispatch);
+ const mapDispatchToProps  = dispatch => bindActionCreators({ sendMessage }, dispatch);
  
  export default connect(mapStateToProps , mapDispatchToProps)(MessageField);
