+++ static_src/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import chatReducer from './chatReducer';
+import messageReducer from './messageReducer';
 
 export default combineReducers({
     chatReducer,
+    messageReducer,
 });
